Add generalized kademlia structures.
This implementation contains only the table itself, the distance function, and a little bit of machinery for adding peers to the table. The relevant properties of the peers in the table are captured in a few traits.

Type `coop.rchain.kademlia.PeerTable` is parameterized on a `Peer` subtype, which provides the bare minimum that has to be implemented by the user: a key and a `ping` method. The `ping` method isn't fully fleshed out; it should eventually have some relationship with latency and potentially trigger an update of the kademlia table, and that requires a synchronous effect (measurement.)

I've included a stub for the rchain usage of kademlia.